[RFC] gdb: introduce helper functions to support gdb debugging
Introduce helper functions to make it easy to find the load addresses
and the TA about to be loaded from GDB.

Without something like this, it's more complicated and error prone to
find the correct information. Without it one has to look at the stack
frames and draw conclusions based on that.

Signed-off-by: Joakim Bech <joakim.bech@linaro.org>